Create setting for audible tab query, somewhat significant refactor
A callback function was implemented to get all extension options when refreshing tabs; the options are collected, and a callback function is invoked with the options passed as a parameter, which does the work of querying and refreshing tabs based on the options passed.
Functions were made more readable and parameters to chrome api functions were declared as variables.