corrected mis-implementation of ignore audible tabs
Refreshing only tabs that are audible is not a feature of the chrome api, but rather this extension.
Therefore, it was necessary to correct the logic such that audible tabs are not queried (which would previously target ONLY audible tabs if the option was set to true).

Instead, if the refreshAudibleTabs setting is set to false AND a queried tab has the audible attribute, it will not be refreshed by this extension.